+++ lib/open_feature/no_op_provider.rb
@@ -5,10 +5,8 @@ module OpenFeature
   # Default provider when initializing OpenFeature.
   # Always returns the default value given.
   # This will result in a TypeError if the given default value does not have the correct type.
-  class NoOpProvider
+  class NoOpProvider < Provider
     extend T::Sig
-
-    include Provider
 
     sig { override.returns(ProviderMetadata) }
     attr_reader :metadata
@@ -20,6 +18,7 @@ module OpenFeature
     def initialize
       @metadata = T.let(ProviderMetadata.new(name: "No Op Provider"), ProviderMetadata)
       @hooks = T.let([], T::Array[Hook])
+      super()
     end
 
     sig do