+++ test/open_feature/configuration_test.rb
@@ -2,8 +2,6 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require_relative "../support/test_provider"
-require_relative "../support/test_hook"
 
 class ConfigurationTest < Minitest::Test
   def teardown